% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chem_sub_extract.R
\name{chem_sub_extract}
\alias{chem_sub_extract}
\title{Easy helper for 'chem_sub_extract'}
\usage{
chem_sub_extract(con, schema, table, time_frame = c("FY", "Monthly"))
}
\arguments{
\item{con}{The database connection object to be used}

\item{schema}{The scheme name to extract data from}

\item{table}{The fact table name to extract data from}

\item{time_frame}{"FY"/"Monthly" - the time frame you which to summarise to}
}
\description{
Extract chemical substance level table from HRT fact table
}
