% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/national_extract.R
\name{national_extract}
\alias{national_extract}
\title{Easy helper for 'national_extract'}
\usage{
national_extract(con, table = "HRT_FACT_DIM", time_frame = c("FY", "Monthly"))
}
\arguments{
\item{con}{The database connection object to be used}

\item{table}{The fact table name to extract data from (defaults to HRT_FACT_DIM)}

\item{time_frame}{"FY"/"Monthly" - the time frame you which to summarise to}
}
\description{
Extract national level table from HRT fact table
}
