% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ons_stp_pop.R
\name{ons_stp_pop}
\alias{ons_stp_pop}
\title{STP population reference file import function}
\source{
\url{https://www.ons.gov.uk/peoplepopulationandcommunity/populationandmigration/populationestimates/datasets/clinicalcommissioninggroupmidyearpopulationestimates}
}
\usage{
ons_stp_pop()
}
\description{
This function downloads mid-year 2020 STP population estimates from ONS, joins 
the data with the ONS lookup and returns population totals for each STP.
}
\examples{
ons_stp_pop()

}
