% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/per_s_error.R
\name{per_s_error}
\alias{per_s_error}
\title{Easy helper for 'per_s_error'}
\usage{
per_s_error(base1, per1, base2, per2)
}
\arguments{
\item{base1}{the base size of the first sample}

\item{per1}{the proportion of the first sample to test}

\item{base2}{the base size of the second sample}

\item{per2}{the proportion of the second sample to test}
}
\description{
Test for significant change between 2 proportions when bases are known
}
