% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quintile_extract.R
\name{quintile_extract}
\alias{quintile_extract}
\title{Easy helper for 'quintile_extract'}
\usage{
quintile_extract(con, table = "HRT_FACT_DIM", time_frame = c("FY", "Monthly"))
}
\arguments{
\item{con}{The database connection object to be used}

\item{table}{The fact table name to extract data from (defaults to HRT_FACT_DIM)}

\item{time_frame}{"FY"/"Monthly" - the time frame you which to summarise to}
}
\description{
Extract imd quintile table from HRT fact table
}
