% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_extract_excel.R
\name{pi_extract_excel}
\alias{pi_extract_excel}
\title{Easy helper for 'pi_extract_excel'}
\usage{
pi_extract_excel(con, schema, table, time_frame = c("FY", "Monthly"))
}
\arguments{
\item{con}{The database connection object to be used}

\item{schema}{The scheme name to extract data from}

\item{table}{The fact table name to extract data from}

\item{time_frame}{"FY"/"Monthly" - the time frame you which to summarise to}
}
\description{
Extract patient identification rate long table from HRT fact table
}
